Skills Suggestion moved as simplified UI
