+++ web/components/Form/Account/SkillsAddForm/SkillsAddForm.tsx
@@ -15,7 +15,7 @@ import { array, object, SchemaOf, string } from 'yup'
 import { yupResolver } from '@hookform/resolvers/yup'
 import { Props, RefHandler } from './types'
 import CreatableAutocomplete from '../../CreatableAutocomplete/CreatableAutocomplete'
-import SkillsSuggest from '@/components/Account/SkillsSuggest/SkillsSuggest'
+import SkillsSuggest from '@/components/UI/SkillsSuggest/SkillsSuggest'
 import Icon from '@/components/Icon/Icon'
 import Button from '@/components/UI/Button/Button'
 import { fetchMe } from '@/service/me'
